requirements.txt
aiohttp==3.8.1
aiohttp-socks==0.7.1
aiorpcX==0.18.7
aiosignal==1.2.0
altgraph==0.17.2
anyio==3.5.0
apt-clone==0.2.1
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.1
attrs==21.2.0
auto-py-to-exe==2.15.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bitarray==2.4.0
bitstring==3.1.9
bleach==4.1.0
blinker==1.4
bottle==0.12.19
bottle-websocket==0.2.9
Brlapi==0.7.0
capstone==5.0.0rc2
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
Click==7.0
colorama==0.4.4
colored-traceback==0.3.0
command-not-found==0.3
configobj==5.0.6
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
dnspython==2.0.0
docopt==0.6.2
dploy==0.1.2
Eel==0.12.4
egcd==0.2.0
Electrum==4.1.5
entrypoints==0.3
evdev==1.4.0
executing==0.8.3
frozenlist==1.2.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
gitdb==4.0.9
GitPython==3.1.27
gmpy==1.17
gmpy2==2.1.0b5
greenlet==1.1.2
grpcio==1.16.1
helpdev==0.7.1
httplib2==0.14.0
idna==2.8
ifaddr==0.1.6
IMDbPY==6.8
importlib-resources==5.4.0
intervaltree==3.1.0
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==2.10.1
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.13.5
jupyterlab==3.3.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
kazam==1.4.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libevdev==0.5
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib-inline==0.1.3
meson==0.53.2
mistune==0.8.4
mpmath==1.2.1
multidict==5.2.0
nbclassic==0.3.6
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nemo-emblems==5.0.0
nest-asyncio==1.5.4
netaddr==0.7.19
netifaces==0.10.4
notebook==6.4.8
notebook-shim==0.1.0
oauthlib==3.1.0
onboard==1.4.1
packaging==20.3
PAM==0.4.2
pandocfilters==1.5.0
paramiko==2.10.3
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipreqs==0.4.11
plumbum==1.7.2
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.1
proton-client==0.7.1
protonvpn-cli==3.11.0
protonvpn-gui==1.7.0
protonvpn-nm-lib==3.7.0
psutil==5.5.1
ptyprocess==0.7.0
pure-eval==0.2.2
pwn==1.0
pwntools==4.7.0
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.10.4
pycups==1.9.73
pycurl==7.43.0.2
pyelftools==0.28
Pygments==2.11.2
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.5
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynput==1.7.6
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyparted==3.11.2
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-gnupg==0.4.5
python-magic==0.4.16
python-socks==2.0.0
python-xapp==2.2.1
python-xlib==0.23
pythondialog==3.4.0
pytz==2019.3
pyudev==0.21.0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
QDarkStyle==2.8.1
qrcode==7.3.1
QtPy==1.11.3
r2env==0.5.0
reportlab==3.5.34
requests==2.25.1
requests-file==1.4.3
requests-unixsocket==0.2.0
ROPGadget==6.7
rpyc==5.1.0
SecretStorage==2.3.1
Send2Trash==1.8.0
setproctitle==1.1.10
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==1.9.5
speedtest-cli==2.1.2
ssh-import-id==5.10
stack-data==0.2.0
sympy==1.3
systemd-python==234
terminado==0.13.2
testpath==0.6.0
tinycss2==1.0.2
tldextract==2.2.1
toml==0.10.2
tornado==6.1
tqdm==4.63.1
traitlets==5.1.1
typing-extensions==4.0.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unicorn==2.0.0rc6
Unidecode==1.1.1
urllib3==1.26.5
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
wget==3.2
whichcraft==0.6.1
xkit==0.0.0
yarg==0.1.9
yarl==1.7.2
youtube-dl==2021.4.26
z3-solver==4.8.15.0
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
zxing==0.14
